# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bitbiz_hias-core_domains_policy_schema.BulkIDsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.ChangePlanRequest:
    properties:
      new_plan_id:
        type: string
      reason:
        type: string
    required:
    - new_plan_id
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.CreatePolicyRequest:
    properties:
      end_date:
        type: string
      plan_id:
        type: string
      policyholder_email:
        type: string
      policyholder_name:
        type: string
      policyholder_phone:
        type: string
      start_date:
        type: string
    required:
    - plan_id
    - policyholder_email
    - policyholder_name
    - policyholder_phone
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.UpdatePolicyRequest:
    properties:
      end_date:
        type: string
      policyholder_email:
        type: string
      policyholder_name:
        type: string
      policyholder_phone:
        type: string
      start_date:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Health Insurance Administration System - Core API
  title: HIAS Core API
  version: "1.0"
paths:
...
  /api/v1/policies:
    get:
      description: List all policies with pagination and optional date/search filters
      parameters:
      - description: Filter from date (RFC3339 or YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter to date (RFC3339 or YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Search by policy number, holder name, or email
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List policies
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Create a new insurance policy
      parameters:
      - description: Policy creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new policy
      tags:
      - Policies
  /api/v1/policies/{id}:
    get:
      description: Retrieve a single policy by its unique identifier
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a policy by ID
      tags:
      - Policies
    put:
      consumes:
      - application/json
      description: Update an existing insurance policy by its ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Policy update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.UpdatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a policy
      tags:
      - Policies
  /api/v1/policies/{id}/activate:
    put:
      description: Activate an insurance policy by its ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Activate a policy
      tags:
      - Policies
...
  /api/v1/policies/{id}/change-plan:
    put:
      consumes:
      - application/json
      description: Change the plan associated with an insurance policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: New plan details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.ChangePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change policy plan
      tags:
      - Policies
...
  /api/v1/policies/{id}/lapse:
    put:
      description: Mark a policy as lapsed due to non-payment or expiry
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lapse a policy
      tags:
      - Policies
...
  /api/v1/policies/{id}/prorate:
    get:
      description: Calculate the prorated premium for a policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Calculate prorated premium
      tags:
      - Policies
...
  /api/v1/policies/{id}/reinstate:
    put:
      description: Reinstate a previously lapsed or suspended policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reinstate a policy
      tags:
      - Policies
...
  /api/v1/policies/{id}/suspend:
    put:
      description: Temporarily suspend an insurance policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Suspend a policy
      tags:
      - Policies
  /api/v1/policies/{id}/terminate:
    put:
      description: Permanently terminate an insurance policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Terminate a policy
      tags:
      - Policies
...
  /api/v1/policies/bulk/activate:
    post:
      consumes:
      - application/json
      description: Activate multiple policies in a single request
      parameters:
      - description: List of policy IDs to activate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.BulkIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk activate policies
      tags:
      - Policies
  /api/v1/policies/bulk/lapse:
    post:
      consumes:
      - application/json
      description: Lapse multiple policies in a single request
      parameters:
      - description: List of policy IDs to lapse
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.BulkIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk lapse policies
      tags:
      - Policies
  /api/v1/policies/by-status:
    get:
      description: List policies filtered by their status
      parameters:
      - description: Policy status to filter by
        in: query
        name: status
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List policies by status
      tags:
      - Policies
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"